allow keyframe interval to be calculated depending on the source
added a new groupbox called video options which contain two comboboxes to select the source framerate and the desired keyframe interval in seconds. from that information, the keyframe interval is calculated in frames by multiplying the two for the video encoders. this certainly makes my life better :D

also do a little bit of cleanup here and there