sonarcloud's recommendations turn things from bad to worse
decided to just make an immutable dictionary

a nuance to this was that tests failed because when you convert a dictionary to immutable, it doesn't preserve the order of the key-value pairs. odd, very odd. so i decided not to use the immutable dictionaries to get the keys and kept the original implementation for that