re-license under the MIT license
when i was studying for the az-400, i had learned that open source licenses fall into two distinct categories, copyleft and permissive

gpl is copyleft, any derivative work must be licensed the same and be open source. this is also known as a viral license. if you happen to reference a package in your software that has a viral license and you not licensing your software the same or it ain't open source, it can cause serious legal problems

the mit license is a permissive license, allowing the use of the software or its components to be a part of proprietary solutions and has no restrictions on how derivative work is licensed

the mit license also doesn't require a header in all source files